proper testing
