+++ JavaScript/tsconfig.json
@@ -18,7 +18,7 @@
       "skipLibCheck": true,
       "outDir": "dist"
     },
-    "include": ["src"],
+    "include": ["src", "test"],
     "exclude": ["node_modules"]
   }
